pyproject.toml
[tool.poetry]
name = "budget-planner"
version = "0.1.0"
description = "Budget planner"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.6"
django = "^5.1.3"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.10"
pre-commit = "^4.0.1"
django-crispy-forms = "^2.3"
crispy-bootstrap4 = "^2024.10"
coverage = "^7.6.9"
django-rest-framework = "^0.1.0"
django-extensions = "^3.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
